# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrainingLog API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/training/api"
    }
  ],
  "security": [
    {
      "ownerCookie": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ownerCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TrainingLog.Auth"
      },
      "watchBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped device token. Cookie authentication is not accepted on watch/v1."
      }
    },
    "schemas": {
...
      "WatchSnapshot": {
        "type": "object",
        "properties": {
          "protocolVersion": {
            "const": 1,
            "type": "integer"
          },
          "contractVersion": {
            "const": 2,
            "type": "integer"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "generation": {
            "type": "string"
          },
          "control": {
            "$ref": "#/components/schemas/WatchControl"
          },
          "payload": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "protocolVersion",
          "contractVersion",
          "sessionId",
          "version",
          "generation",
          "control",
          "payload"
        ]
      },
...
      "WatchPairing": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[0-9]{8}$"
          },
          "expiresAt": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "code",
          "expiresAt"
        ]
      },
      "WatchRedeem": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[0-9]{8}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "WatchCredentials": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer",
            "minimum": 0
          },
          "protocolVersion": {
            "type": "integer",
            "const": 1
          },
          "contractVersion": {
            "type": "integer",
            "const": 2
          }
        },
        "required": [
          "deviceId",
          "token",
          "expiresAt",
          "protocolVersion",
          "contractVersion"
        ]
      },
      "WatchDevice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer",
            "minimum": 0
          },
          "revokedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "lastSeenAt": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "name",
          "expiresAt",
          "revokedAt",
          "lastSeenAt"
        ]
      },
...
    }
  },
  "paths": {
...
    "/watch/pairing": {
      "post": {
        "tags": [
          "Apple Watch"
        ],
        "security": [
          {
            "ownerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchPairing"
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "postWatchPairing"
      }
    },
    "/watch/pairing/redeem": {
      "post": {
        "tags": [
          "Apple Watch"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchRedeem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchCredentials"
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "postWatchPairingRedeem"
      }
    },
    "/watch/devices": {
      "get": {
        "tags": [
          "Apple Watch"
        ],
        "security": [
          {
            "ownerCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "devices": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WatchDevice"
                      }
                    }
                  },
                  "required": [
                    "enabled",
                    "devices"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "getWatchDevices"
      }
    },
    "/watch/devices/{id}": {
      "delete": {
        "tags": [
          "Apple Watch"
        ],
        "security": [
          {
            "ownerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "revoked"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "deleteWatchDevicesId"
      }
    },
    "/sessions/{id}/watch-control": {
      "get": {
        "tags": [
          "Apple Watch"
        ],
        "security": [
          {
            "ownerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "getSessionsIdWatchControl"
      }
    },
...
    "/sessions/{id}/watch-handoff/cancel": {
      "post": {
        "tags": [
          "Apple Watch"
        ],
        "security": [
          {
            "ownerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "postSessionsIdWatchHandoffCancel"
      }
    },
...
    "/watch/recovery": {
      "get": {
        "tags": [
          "Apple Watch"
        ],
        "security": [
          {
            "ownerCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WatchRecovery"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "401": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "409": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "423": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "426": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "429": {
            "description": "Authentication, validation, version or control error; local records must be retained."
          },
          "503": {
            "description": "New watch pairing or handoff is disabled; existing control remains enforced."
          }
        },
        "operationId": "getWatchRecovery"
      }
    },
...
  }
}